Rotation Work, and Something Else...
So, you see, I had an issue. I decided to use .unless while creating a command, hoping that this function would only do something if a condition returned false. So, if this number is -1, do this. Instead, .unless reads "If this number is not -1, do this". So, I fix that. :D I also added something else. Don't mind that. Not important.

- Aidan