Pathfinding Work
I still want to test the pathfinding we had, but I have created a set of paths and controls for the robot to use. It will pathfind to the starting position of the path, and then run the path, hopefully resulting in more accurate alignment.
Also cleaned up some other odd things that we don't mention to the drive team.