6 Coral Auto, Pathfinding Improvements
Yesterday I made the auto fast enough to achieve 6 L4 coral in under 15 seconds, able to pickup another coral during that time. Along with this, I've made some improvements to the pathfinding (or rather, completely refactored it). This results in what I would consider 10 times more reliable paths, except in rotation. I intend to use this concept in teaching a few other students the programming, while also improving upon the concept, giving them a chance to work on the actual robot. Therefore, these changes will not come to the robot for some time.