Alliance Pathfinding + Better Positions
We are back for offseason, and will start working on optimizing much of our robot, both to prepare for our offseason competitions, and to train for next season. Previously, I, and one other, worked on revising pathfinding. We found issues with how to change pathfinding options based on alliance. This is due to how a command is initialized, and we were giving it information on initialization, meaning that the information stayed for the rest of the time. This release puts everything into the execute() method, meaning the command will change the two pose arrays correctly.