Error Checking
I think we will be testing pathfinding on the actual robot tomorrow! However, before we do, I want to ensure that nothing can give out an error easily. This includes making sure an alliance is set before checking which alliance is set in the pathfinding code. This change shouldn't mean much, but it is a nice safeguard. Also removed some telemetry stuff used from the simulation, which I don't want to have on the actual robot, to free up resources as much as possible.