Elevator State Machine Testing
One thing I want to incorporate into our robot code during offseason is a command-based state machine. The robot subsystems all have states storing different information for that subsystem, and this means we can sequence actions using triggers. I'm starting this with the elevator, and it runs the same as before, but will now be able to easier communicate with the rest of the robot. Also, removed the non-simulator ElevatorSubsystem, as it was just taking memory without any use.