Auto Testing
I wanted to see what we need for a 4 coral auto, as this is one of our goals for the season, so I've been testing in simulation with the faster elevator speed we hope to achieve on the actual robot. The goal we have, now looking, is absolutely ridiculous with what we have currently. I think a ground intake may be required for what we hope to do. I also added Elevator States to the second player robot, and made that elevator faster. There is also a elevator slow mode state, which will be active when a path is running that may be a concern for this. The ground state doesn't really need this mode, but I've added it anyways.