Effector State Machine + Auto Testing
Currently, we have finished a basic state machine for the end effector, which makes much of our other code (located in full classes) compress to about four lines. The effector state can be set in just one line of code, affecting the wheels and conveyor. The only thing that doesn't have a state model is the affector pivot. Autos also aren't working fully, due to the affector pivot. (Honestly might be a good thing in this case)